Creating a Volcano plot

It performs a Volcano plot on the results of statistical analyses such as t-test and a Fold Change.

Usage

GCP_Volcano(
  GCPlist,
  x_val = "logFC",
  y_val = "ttest_PvaluesFDR",
  log_base = getOption("GetCoolProteopipe.log_base"),
  pcutoff_colored = 0.05,
  pcutoff_line = 0.05,
  pcutoff_prot_label = NULL,
  name_column_proteinlabels = NULL,
  name_column_proteingroups = NULL,
  col_pal_difference = c("grey", "blue", "red"),
  col_pal_groups = NULL
)

Arguments

GCPlist

a list initially created with ImportOutputMaxQuan(), ImportPTMs(), or ImportOutputProtDiscov().

x_val

character of length 1. A column name of the ProteinINFO table, containing values that will be used for the x-axis. Typically the "logFC" column, that is generated after the function GCP_FoldChange, or one of the columns generated by the function GCP_FoldChange_Multi, for example: "gr1_vs_gr2_logFC".

y_val

character of length 1. A column name of the ProteinINFO table, containing values that will be used for the y-axis, after calculating the -log10 of them. Pass here a column containing p-values, such as "ttest_PvaluesFDR", or "ttest_Pvalues", or "ANOVA_gr2_vs_gr1_Pvalue", or "ANOVA_gr2_vs_gr1_PvalueFDR".

log_base

numeric of length 1. The log-base passed to GCP_FoldChange.

pcutoff_colored

NULL or a numeric of length 1. The p-value cut-off below which showing colored dots. If NULL, dots will not be colored by significance.

pcutoff_line

NULL or a numeric of length 1. The p-value where a horizontal dahsed line will be added to the plot. If NULL, this line will not be showed.

pcutoff_prot_label

NULL or a numeric of length 1. The p-value below which protein names will be added to the plot. If NULL, no names will be reported.

name_column_proteinlabels

NULL or a character of length 1. The name of a column in the proteinINFO table containing the names you would plot.

name_column_proteingroups

NULL or a character of length 1. The name of a column in the proteinINFO table containing groups of protein. If you pass this argument, the dots will be colored based on these groups, and this will replace whether you specify pcutoff_colored.

col_pal_difference

a character vector containing exactly 3 colors. If pcutoff_colored is specified, they will be used for Not significant, Down, and Up proteins, respectively.

col_pal_groups

NULL or a character vector with colors to be used if name_column_proteingroups were specified. If NULL, colors from the pals package will be used (see function GetFeatistics::build_long_vector_of_colors).

Value

A ggplot object.

Examples

if (FALSE) { # \dontrun{

Fig15_Volcano_plot <- GCP_Volcano(GCPlist = GCPlist14,
                                  x_val = "logFC",
                                  y_val = "ttest_PvaluesFDR",
                                  pcutoff_colored = 0.05,
                                  pcutoff_line = 0.05,
                                  pcutoff_prot_label = 0.005,
                                  name_column_proteinlabels = "Protein names",
                                  name_column_proteingroups = NULL,
                                  col_pal_difference = c("grey", "blue", "red"),
                                  col_pal_groups = NULL)
export_figures(Fig15_Volcano_plot)

} # }